Get MSBuildWorkspace tests running again
This change moves the MSBuildWorkspace tests to a separate project and introduces infrastructure to get them
running. Essentially, each test is marked with a `ConditionalFact` with `VisualStudioMSBuildInstalled` as the
skip condition. That skip condition uses the VS 2017 setup APIs to locate installed instances of VS and finds
one that contains the Roslyn Compiler package. Then, it sets up an AssemblyResolve event on the current AppDomain
to resolve MSBuild locations from VS itself. This ensures that MSBuild can locate a toolset containing the
necessary props and targets to handle the projects in our tests.