Factor IEqualityComparer<T> for ImmutableHashSet into a utility class
CompletionServiceWithProviders directly implemented IEqualityComparer<ImmutableHashSet<string>>
to handle comparison between hash sets for text view roles. This change removes that implementation
and factors it into a utility class to allow it to be used by other similar features (e.g. Signature
Help). This is technically a breaking change, but it's very unlikely that any consumers of
CompletionService would depend on the fact that one of its subtypes implemented
IEqualityComparer<ImmutableHashSet<string>>.